Revert "Epoch time control"
